+++ Runtime/Ponito/Core/Asyncs/Extensions/Exts.cs
@@ -1,57 +1,65 @@
-using System;
 using System.Collections;
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 using Ponito.Core.Asyncs.Interfaces;
 using Ponito.Core.Asyncs.Promises;
 using Ponito.Core.Asyncs.Tasks;
-using Ponito.Core.Asyncs.Tasks.Movables;
-using Ponito.Core.DebugHelper;
 using UnityEngine;
 
 namespace Ponito.Core.Asyncs.Extensions
 {
     public static partial class Exts
     {
-        private static YieldInstruction Eof = new WaitForEndOfFrame();
+        private static readonly YieldInstruction eof = new WaitForEndOfFrame();
 
         [DebuggerHidden]
-        public static Movable<T> AsPoTask<T>(this ValueTask<T> vt)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static async PoTask AsPoTask(this ValueTask vt)
         {
-            return new ValueTaskAwait<T>(vt);
+            while (!vt.IsCompleted) await PoTask.Yield();
         }
 
         [DebuggerHidden]
-        public static Movable AsPoTask(this ValueTask vt)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static async PoTask<T> AsPoTask<T>(this ValueTask<T> vt)
         {
-            return new ValueTaskAwait(vt);
+            while (!vt.IsCompleted) await PoTask.Yield();
+            return vt.Result;
         }
 
         [DebuggerHidden]
-        public static Movable<T> AsPoTask<T>(this Task<T> t)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static async PoTask AsPoTask(this Task t)
         {
-            return new TaskAwait<T>(t);
+            while (!t.IsCompleted) await PoTask.Yield();
         }
 
         [DebuggerHidden]
-        public static Movable AsPoTask(this Task t)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static async PoTask<T> AsPoTask<T>(this Task<T> t)
         {
-            return new TaskAwait(t);
+            while (!t.IsCompleted) await PoTask.Yield();
+            return t.Result;
         }
 
+        [DebuggerHidden]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static IEnumerator RunAsCoroutine(this Movable movable, Promise p = null)
         {
             if (p != null) p.State = PromiseState.Doing;
-            while (movable.MoveNext()) yield return Eof;
+            while (movable.MoveNext()) yield return eof;
             if (p != null) p.State = PromiseState.Done;
         }
 
+        [DebuggerHidden]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static IEnumerator RunAsCoroutine(this PoTask task, Promise p = null)
         {
             var movable            = task.GetAwaiter();
             if (p != null) p.State = PromiseState.Doing;
-            while (movable.MoveNext()) yield return Eof;
+            while (movable.MoveNext()) yield return eof;
 
             if (p != null && movable.Ex != null)
             {
@@ -63,11 +71,13 @@ namespace Ponito.Core.Asyncs.Extensions
             if (p != null) p.State = PromiseState.Done;
         }
 
+        [DebuggerHidden]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static IEnumerator RunAsCoroutine<T>(this PoTask<T> task, Promise<T> p = null)
         {
             var movable            = task.GetAwaiter();
             if (p != null) p.State = PromiseState.Doing;
-            while (movable.MoveNext()) yield return Eof;
+            while (movable.MoveNext()) yield return eof;
 
             if (p != null && movable.Ex != null)
             {