+++ Runtime/Ponito/Core/Asyncs/Tasks/Movables/PredicateAwait.cs
@@ -5,8 +5,9 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
 {
     internal class PredicateAwait : MovableBase
     {
-        public PredicateAwait(Func<bool> predicate, bool waitForTrue = false)
+        public PredicateAwait(Func<bool> predicate, bool waitForTrue = false, CancellationToken ct = default)
         {
+            Ct               = ct;
             this.predicate   = predicate;
             this.waitForTrue = waitForTrue;
         }
@@ -16,8 +17,10 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
 
         public override bool MoveNext()
         {
+            if (Ct.IsCancellationRequested) return false;
             if (IsCompleted) return false;
-            if (waitForTrue == predicate()) return true;
+            if (waitForTrue  && predicate()) return true;
+            if (!waitForTrue && !predicate()) return true;
             return ContinueMoveNext();
         }