+++ Runtime/Ponito/Core/Asyncs/Tasks/Movables/TaskAwait`1.cs
@@ -4,12 +4,12 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
 {
     internal class TaskAwait<T> : MovableBase<T>
     {
-        private Task<T> task { get; }
-
         public TaskAwait(Task<T> task)
         {
             this.task = task;
         }
+
+        private Task<T> task { get; }
 
         public override T GetResult()
         {
@@ -18,6 +18,7 @@ namespace Ponito.Core.Asyncs.Tasks.Movables
 
         public override bool MoveNext()
         {
+            if (Ct.IsCancellationRequested) return false;
             if (IsCompleted) return false;
             if (!task.IsCompleted) return true;
             return ContinueMoveNext();