+++ Runtime/Ponito/Core/Asyncs/Tasks/PoTask.Factory.cs
@@ -11,37 +11,44 @@ namespace Ponito.Core.Asyncs.Tasks
 {
     public partial class PoTask
     {
+        
         [DebuggerHidden]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Movable Yield(CancellationToken ct = default)
         {
             return new YieldAwait(ct);
         }
 
         [DebuggerHidden]
-        public static Movable Delay(int milliseconds)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static Movable Delay(int milliseconds, CancellationToken ct = default)
         {
-            return new DelayAwait(milliseconds);
+            return new DelayAwait(milliseconds, ct);
         }
 
         [DebuggerHidden]
-        public static Movable Delay(float seconds)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static Movable Delay(float seconds, CancellationToken ct = default)
         {
-            return new DelayAwait(seconds);
+            return new DelayAwait(seconds, ct);
         }
 
         [DebuggerHidden]
-        public static Movable WaitWhile(Func<bool> predicate)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static Movable WaitWhile(Func<bool> predicate, CancellationToken ct = default)
         {
-            return new PredicateAwait(predicate, true);
+            return new PredicateAwait(predicate, true, ct);
         }
 
         [DebuggerHidden]
-        public static Movable WaitUntil(Func<bool> predicate)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static Movable WaitUntil(Func<bool> predicate, CancellationToken ct = default)
         {
-            return new PredicateAwait(predicate);
+            return new PredicateAwait(predicate, false, ct);
         }
 
         [DebuggerHidden]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static Movable Create(IEnumerator ie, CancellationToken ct = default)
         {
             return new IEnumeratorAwait(ie, ct);